Agent: Move IntRange into common.types
