+++ monkey/common/types/__init__.py
@@ -1,6 +1,7 @@
 from .concurrency import Lock, Event
 from .serialization import JSONSerializable
 from .ids import AgentID, HardwareID, MachineID
+from .int_range import IntRange
 from .networking import NetworkService, NetworkPort, PortStatus, SocketAddress
 from .plugin_types import PluginName
 from .plugin_types import PluginVersion
